app/requirements.txt
boto3
Django==1.11.20
django-autocomplete-light==3.5.1
django-bootstrap3==9.0.0 # TODO delete after removing all template extensions
django-bootstrap-datepicker-plus==3.0.5
django-countries==5.1.1
django-health-check==3.7.0
django-jquery==3.1.0
django-stronghold==0.2.8
djangorestframework==3.9.1
djangorestframework-jwt==1.9.0
django-smtp-ssl==1.0
furl==1.0.1
mock==2.0.0
mysqlclient==1.3.13
Pillow==6.2.0
py-auth0-jwt==0.3.0
py-auth0-jwt-rest==0.1.2
python-pstore==0.8
PyJWT==1.6.1
PyMySQL==0.7.9
raven==6.1.0
requests==2.20.0
sqlparse==0.2.4
